refactor: move source files to `k-event-manager` module
- Renamed package from `com.fantamomo.kevent.impl` to `com.fantamomo.kevent.manager`.
- Moved source files to `k-event-manager/src/commonMain/kotlin` for improved modularity and alignment with project structure.
