+++ k-event-manager/src/commonMain/kotlin/com/fantamomo/kevent/manager/DefaultEventManager.kt
@@ -1,3 +1,3 @@
-package com.fantamomo.kevent.impl
+package com.fantamomo.kevent.manager
 
 class DefaultEventManager : AbstractEventManager()
