+++ k-event-manager/src/commonMain/kotlin/com/fantamomo/kevent/manager/EventManager.kt
@@ -1,4 +1,4 @@
-package com.fantamomo.kevent
+package com.fantamomo.kevent.manager
 
 import kotlin.reflect.KClass
